requirements.txt
# ============================================================
#  ITHACA-FV Agentic RAG — Python Dependencies
#  Install with:  pip install -r requirements.txt
# ============================================================

# --- Core RAG / LLM framework ---
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langgraph>=0.1.0           # agentic workflow graph

# --- LLM provider integrations ---
langchain-openai>=0.1.0    # used for OpenRouter + vLLM (both are OpenAI-compatible)

# --- Local embeddings ---
sentence-transformers>=3.0.0
torch>=2.2.0               # needed by sentence-transformers

# --- Vector store ---
chromadb>=0.5.0            # local persistent vector DB

# --- Document loading ---
gitpython>=3.1.40          # clone / read GitHub repos
unstructured>=0.12.0       # generic file parser
markdown>=3.5              # Markdown parsing

# --- Configuration ---
pyyaml>=6.0
python-dotenv>=1.0.0

# --- GUI ---
gradio>=4.30.0

# --- Utilities ---
tqdm>=4.66.0               # progress bars during ingestion
loguru>=0.7.0              # structured logging
rich>=13.0.0               # pretty console output

# --- Dev / testing ---
pytest>=8.0.0
pytest-asyncio>=0.23.0
